+++ assets/js/src/bulk-delete.js
@@ -133,4 +133,12 @@ jQuery(document).ready(function () {
 			return false;
 		}
 	};
+
+    BulkWP.validateUserMeta = function() {
+        if (jQuery('#smbd_u_meta_value').val() !== '') {
+            return true;
+        } else {
+            return false;
+        }
+    };
 });